refactor(examples): CLI対戦例の相手をjpoke提供のMaxDamagePlayerに変更
自作方策を再実装せず、jpoke.players.MaxDamagePlayerをそのまま対戦相手として
使うよう簡素化。実体が自作AIではなくなったためファイル名も05_cli_vs_ai.pyに変更。
